requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cmake==4.1.0
colorama==0.4.6
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
executing==2.2.1
fastapi==0.119.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
ipykernel==7.0.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain==0.3.27
langchain-core==0.3.79
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langgraph==1.0.0a4
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.7.0a2
langgraph-sdk==0.2.9
langsmith==0.4.34
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
openai==2.3.0
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
parso==0.8.5
platformdirs==4.5.0
prompt_toolkit==3.0.52
psutil==7.1.0
pure_eval==0.2.3
pydantic==2.12.2
pydantic_core==2.41.4
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.0
sentry-sdk==2.41.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.48.0
tenacity==9.1.2
tiktoken==0.12.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
wheel==0.45.1
xxhash==3.6.0
zstandard==0.25.0